Fix notices on variations without prices (https://github.com/woocommerce/woocommerce-blocks/pull/2722)
* ensure response of `wc_format_decimal` is a float before doing calculating

* Fix notice when empty array passed into min/max

+++ plugins/woocommerce-blocks/src/StoreApi/Schemas/AbstractSchema.php
@@ -240,7 +240,7 @@ abstract class AbstractSchema {
 	protected function prepare_money_response( $amount, $decimals = 2, $rounding_mode = PHP_ROUND_HALF_UP ) {
 		return (string) intval(
 			round(
-				wc_format_decimal( $amount ) * ( 10 ** $decimals ),
+				( (float) wc_format_decimal( $amount ) ) * ( 10 ** $decimals ),
 				0,
 				absint( $rounding_mode )
 			)
